pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sc-sim-tools"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "dagster>=1.7,<2.0",
    "dagster-webserver>=1.7,<2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.dagster]
module_name = "sc_sim_pipeline"
